Extract badge component out of proposals
