+++ src/app/components/Proposals/ProposalStatusIcon.tsx
@@ -1,31 +1,11 @@
 import { FC } from 'react'
 import { TFunction } from 'i18next'
 import { useTranslation } from 'react-i18next'
-import Box from '@mui/material/Box'
 import CheckCircleIcon from '@mui/icons-material/CheckCircle'
 import CancelIcon from '@mui/icons-material/Cancel'
-import { styled } from '@mui/material/styles'
 import { COLORS } from '../../../styles/theme/colors'
 import { ProposalState } from '../../../oasis-nexus/api'
-
-const StyledBox = styled(Box)(({ theme }) => ({
-  display: 'inline-flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  gap: 3,
-  flex: 1,
-  borderRadius: 10,
-  padding: theme.spacing(2, 2, 2, '10px'),
-  fontSize: '12px',
-  minWidth: '85px',
-}))
-
-const StyledIcon = styled(Box)({
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  fontSize: '18px',
-})
+import { Badge } from '../../components/Badge'
 
 const getStatuses = (t: TFunction) => ({
   [ProposalState.active]: {
@@ -73,11 +53,11 @@ export const ProposalStatusIcon: FC<ProposalStatusIconProps> = ({ status }) => {
   const IconComponent = statusConfig.icon
 
   return (
-    <StyledBox sx={{ backgroundColor: statusConfig.backgroundColor, color: statusConfig.textColor }}>
-      {statusConfig.label}
-      <StyledIcon sx={{ color: statusConfig.iconColor }}>
-        <IconComponent color="inherit" fontSize="inherit" />
-      </StyledIcon>
-    </StyledBox>
+    <Badge
+      backgroundColor={statusConfig.backgroundColor}
+      textColor={statusConfig.textColor}
+      label={statusConfig.label}
+      icon={<IconComponent sx={{ color: statusConfig.iconColor }} fontSize="inherit" />}
+    />
   )
 }
